Add files via upload
# Power-bi-projects

## Problem Statement

The current dashboard is designed to analyze and visualize road accident cases in London, using data from the years 2021 and 2022. It serves as a critical tool for understanding accident trends, identifying high-risk areas, and aiding in the development of strategies to improve road safety. While the dashboard provides basic functionalities, there is a need to enhance its capabilities to ensure it delivers more comprehensive and actionable insights.

## Columns that I have used

  (a) Accident Date

  (b)Junction Control
  
  (c) Accident severity
  
  (d) Light Condition
  
  (e) Weather Condition
  
  (f) Vehical Type